# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
      "ActionItemListOut": {
        "description": "The scan's fix list, plus what happened the last time it was generated.",
        "properties": {
          "generatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedby"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "reasonCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoncode"
          },
          "scanId": {
            "title": "Scanid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "scanId",
          "status"
        ],
        "title": "ActionItemListOut",
        "type": "object"
      },
      "ActionItemOut": {
        "description": "One generated fix \u2014 \u00a75.3's prioritised action list.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "dimensionKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensionkey"
          },
          "effort": {
            "$ref": "#/components/schemas/Effort"
          },
          "generatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedby"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pointsUpside": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointsupside"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "scanId": {
            "title": "Scanid",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ActionItemSource"
          },
          "sourceKey": {
            "title": "Sourcekey",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ActionItemStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scanId",
          "source",
          "sourceKey",
          "title",
          "priority",
          "effort",
          "status",
          "rank",
          "createdAt",
          "updatedAt"
        ],
        "title": "ActionItemOut",
        "type": "object"
      },
      "ActionItemSource": {
        "description": "Which Epic 7 candidate produced this fix.\n\nDeliberately only two members. An LLM-authored fix is not a third source \u2014\nit is the same measured gap or audit finding, worded better, and modelling\nit as its own source would let a generated recommendation exist without a\nmeasurement behind it.",
        "enum": [
          "gap",
          "audit"
        ],
        "title": "ActionItemSource",
        "type": "string"
      },
      "ActionItemStatus": {
        "enum": [
          "open",
          "in_progress",
          "done",
          "dismissed"
        ],
        "title": "ActionItemStatus",
        "type": "string"
      },
...
      "ClassificationStatus": {
        "description": "Outcome of Epic 2's industry classification.\n\nDeliberately mirrors Score.status (scoring-spec.md): a result we could not\ndetermine is recorded as such, never as a confident-looking guess.",
        "enum": [
          "pending",
          "classified",
          "ambiguous",
          "unclassifiable"
        ],
        "title": "ClassificationStatus",
        "type": "string"
      },
      "ClientDetailOut": {
        "description": "A client plus the crawl facts from its most recent classification.",
        "properties": {
          "agencyId": {
            "title": "Agencyid",
            "type": "string"
          },
          "brandName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brandname"
          },
          "classificationReasonCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classificationreasoncode"
          },
          "classificationStatus": {
            "$ref": "#/components/schemas/ClassificationStatus"
          },
          "classifiedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifiedat"
          },
          "classifierModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifiermodel"
          },
          "crawl": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrawlSummaryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "industryConfidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industryconfidence"
          },
          "industryConfidenceScore": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industryconfidencescore"
          },
          "industryNiche": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industryniche"
          },
          "kind": {
            "$ref": "#/components/schemas/ClientKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agencyId",
          "name",
          "domain",
          "kind",
          "classificationStatus",
          "createdAt",
          "updatedAt"
        ],
        "title": "ClientDetailOut",
        "type": "object"
      },
      "ClientKind": {
        "description": "A prospect becomes a client when the deal closes.\n\nThe same row transitions rather than being recreated, so the scan history\ngathered while prospecting becomes the \"before\" half of the before/after\nROI report in Epic 11.",
        "enum": [
          "prospect",
          "client",
          "churned"
        ],
        "title": "ClientKind",
        "type": "string"
      },
      "ClientOut": {
        "properties": {
          "agencyId": {
            "title": "Agencyid",
            "type": "string"
          },
          "brandName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brandname"
          },
          "classificationReasonCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classificationreasoncode"
          },
          "classificationStatus": {
            "$ref": "#/components/schemas/ClassificationStatus"
          },
          "classifiedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifiedat"
          },
          "classifierModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifiermodel"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "industryConfidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industryconfidence"
          },
          "industryConfidenceScore": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industryconfidencescore"
          },
          "industryNiche": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industryniche"
          },
          "kind": {
            "$ref": "#/components/schemas/ClientKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agencyId",
          "name",
          "domain",
          "kind",
          "classificationStatus",
          "createdAt",
          "updatedAt"
        ],
        "title": "ClientOut",
        "type": "object"
      },
...
      "CrawlSummaryOut": {
        "description": "FACTS from the crawl. Never page content \u2014 see services/crawl.py.",
        "properties": {
          "detectedLanguage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detectedlanguage"
          },
          "h1Count": {
            "title": "H1Count",
            "type": "integer"
          },
          "hasMetaDescription": {
            "title": "Hasmetadescription",
            "type": "boolean"
          },
          "hasTitle": {
            "title": "Hastitle",
            "type": "boolean"
          },
          "pagesFetched": {
            "title": "Pagesfetched",
            "type": "integer"
          },
          "schemaTypes": {
            "items": {
              "type": "string"
            },
            "title": "Schematypes",
            "type": "array"
          },
          "urlsFetched": {
            "items": {
              "type": "string"
            },
            "title": "Urlsfetched",
            "type": "array"
          },
          "wordCount": {
            "title": "Wordcount",
            "type": "integer"
          }
        },
        "required": [
          "pagesFetched",
          "urlsFetched",
          "schemaTypes",
          "wordCount",
          "h1Count",
          "hasTitle",
          "hasMetaDescription"
        ],
        "title": "CrawlSummaryOut",
        "type": "object"
      },
      "CreateClientRequest": {
        "description": "Intake payload.\n\n`url` is intentionally permissive \u2014 an operator pastes whatever they have,\nincluding a bare domain. Normalisation and validation happen server-side in\nservices/intake.py:parse_domain so the browser and the API cannot disagree\nabout what counts as a valid address.",
        "properties": {
          "classify": {
            "default": true,
            "title": "Classify",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 3,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreateClientRequest",
        "type": "object"
      },
...
      "Effort": {
        "enum": [
          "S",
          "M",
          "L"
        ],
        "title": "Effort",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ClientOut_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ClientOut"
            },
            "title": "Data",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Page[ClientOut]",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Visibility & Competitive Intelligence Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/clients": {
      "get": {
        "description": "Cursor-paginated client list, newest first.\n\nIDs are ULIDs, so `id < cursor ORDER BY id DESC` is a stable, index-backed\nwindow that does not skip or repeat rows when something is inserted\nmid-scroll \u2014 which for an active prospecting list is the common case.",
        "operationId": "list_clients_api_v1_clients_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ClientOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clients",
        "tags": [
          "clients"
        ]
      },
      "post": {
        "description": "Submit a URL: creates the client, crawls it, and classifies its industry.\n\nRuns synchronously. \u00a77's acceptance criterion is a classified industry\nwithin 30 seconds of submitting a URL, and the measured path is a few\nseconds, so one round trip is both sufficient and simpler than polling.\n\nSet `classify: false` to create the record without spending an LLM call.",
        "operationId": "create_client_api_v1_clients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Client",
        "tags": [
          "clients"
        ]
      }
    },
    "/api/v1/clients/{clientId}": {
      "get": {
        "operationId": "get_client_api_v1_clients__clientId__get",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "title": "Clientid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Client",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/api/v1/clients/{clientId}/reclassify": {
      "post": {
        "description": "Re-run crawl and classification for an existing client.\n\nNeeded because a site changes, and because an AMBIGUOUS result is often\nworth retrying after the operator has looked at the site themselves.",
        "operationId": "reclassify_client_api_v1_clients__clientId__reclassify_post",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "title": "Clientid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reclassify Client",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/api/v1/scans/{scanId}/fixes": {
      "get": {
        "description": "The scan's fix list as last generated, in rank order.\n\nReturns an empty list with `status: \"empty\"` rather than 404 when nothing\nhas been generated yet: \"no fixes yet\" is a state of a scan that exists,\nand the report renders its deterministic list in that case.",
        "operationId": "get_fixes_api_v1_scans__scanId__fixes_get",
        "parameters": [
          {
            "in": "path",
            "name": "scanId",
            "required": true,
            "schema": {
              "title": "Scanid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Fixes",
        "tags": [
          "actions"
        ]
      },
      "post": {
        "description": "Generate the scan's prioritised fix list.\n\nMakes **one model call**. The candidates \u2014 which dimension gaps are worth a\nline, which audit findings warrant a fix, and in what order \u2014 are decided\nbefore the call by the same rules Epic 7 already applies on the client. The\nmodel words them and judges priority and effort; it cannot add a\nrecommendation the measurement did not produce.\n\nOne row per candidate per scan, refreshed on re-run. A regenerated list is a\nbetter statement of the same measurement, not a second measurement \u2014 but\n`status` is the operator's, so it survives regeneration, and a candidate\nthat has been worked is kept even once it stops being measured.\n\nA provider failure writes nothing and returns `status: \"failed\"` with a\nreason code. The report still renders its deterministic fix list.",
        "operationId": "generate_fixes_api_v1_scans__scanId__fixes_post",
        "parameters": [
          {
            "in": "path",
            "name": "scanId",
            "required": true,
            "schema": {
              "title": "Scanid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionItemListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Fixes",
        "tags": [
          "actions"
        ]
      }
    },
...
  }
}